backend/api/configuration/serializers/contact.py
from rest_framework import serializers
from configuration.models import Contact

class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = '__all__'
